Move timed_event to event header + cleanup
since the timed_event structure is tightly coupled to the events.c
file, and timed events, we should keep that structure in that module,
so we don't need to include the entire systems data strctures just to
handle some events.

Also, start the cleanup of unused event types, since we don't want
event types at all quite soon, but just callbacks.
